+++ app/Http/Controllers/StudentController.php
@@ -5,34 +5,46 @@ namespace App\Http\Controllers;
 use App\Models\Course;
 use App\Models\Student;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Auth;
 
 class StudentController extends Controller
 {
+    public function __construct()
+    {
+        $this->middleware('auth');
+    }
+
     public function index()
     {
-        $students=Student::latest()->paginate(30);
-        return view('students.index',compact('students'));
+        if (Auth::user()->user_type==1 || Auth::user()->user_type==2) {
+            $students = Student::latest()->paginate(30);
+            return view('students.index', compact('students'));
+        }
     }
 
     public function create()
     {
-        $courses=Course::orderBy('name','ASC')->pluck('name','id');
-        return view('students.create',compact('courses'));
+        if (Auth::user()->user_type==1 || Auth::user()->user_type==2) {
+            $courses = Course::orderBy('name', 'ASC')->pluck('name', 'id');
+            return view('students.create', compact('courses'));
+        }
     }
     public function store(Request $request)
     {
-        $request->validate([
-            'course_id'=>'required',
-            'dni'=>'required|unique:students',
-            'name'=>'required',
-            'last_name'=>'required',
-            'birth_date'=>'required',
-            'phone_number'=>'required|max:9|min:9',
-            'email'=>'required'
-        ]);
+        if (Auth::user()->user_type==1 || Auth::user()->user_type==2) {
+            $request->validate([
+                'course_id' => 'required',
+                'dni' => 'required|unique:students',
+                'name' => 'required',
+                'last_name' => 'required',
+                'birth_date' => 'required',
+                'phone_number' => 'required|max:9|min:9',
+                'email' => 'required'
+            ]);
 
-        Student::create($request->all());
-        return redirect()->route('students.index')->with('success','Se ha creado correctamente');
+            Student::create($request->all());
+            return redirect()->route('students.index')->with('success', 'Se ha creado correctamente');
+        }
     }
 
     /**
@@ -47,30 +59,34 @@ class StudentController extends Controller
     }
     public function edit(Student $student)
     {
-        $courses=Course::orderBy('name','ASC')->pluck('name','id');
-        return view('students.edit',compact('student','courses'));
-
+        if (Auth::user()->user_type==1 || Auth::user()->user_type==2) {
+            $courses = Course::orderBy('name', 'ASC')->pluck('name', 'id');
+            return view('students.edit', compact('student', 'courses'));
+        }
     }
 
     public function update(Request $request, Student $student)
     {
-        $request->validate([
-            'course_id'=>'required',
-            'dni'=>'required|unique:students',
-            'name'=>'required',
-            'last_name'=>'required',
-            'birth_date'=>'required',
-            'phone_number'=>'required|max:9|min:9',
-            'email'=>'required'
-        ]);
-        $student->update($request->all());
-        return redirect()->route('students.index')->with('success','Se ha actualizado');
-
+        if (Auth::user()->user_type==1 || Auth::user()->user_type==2) {
+            $request->validate([
+                'course_id' => 'required',
+                'dni' => 'required|unique:students',
+                'name' => 'required',
+                'last_name' => 'required',
+                'birth_date' => 'required',
+                'phone_number' => 'required|max:9|min:9',
+                'email' => 'required'
+            ]);
+            $student->update($request->all());
+            return redirect()->route('students.index')->with('success', 'Se ha actualizado');
+        }
     }
 
     public function destroy(Student $student)
     {
-        $student->delete();
-        return redirect()->route('students.index')->with('success','Se ha borrado exitosamente');
+        if (Auth::user()->user_type==1 || Auth::user()->user_type==2) {
+            $student->delete();
+            return redirect()->route('students.index')->with('success', 'Se ha borrado exitosamente');
+        }
     }
 }
